+++ packages/toolkit/src/view/catalog/components/FileDetailsOverlay.tsx
