+++ packages/toolkit/src/view/catalog/components/FilePreview.tsx
