+++ packages/toolkit/src/view/catalog/components/MetadataPreview.tsx
