+++ packages/toolkit/src/view/catalog/components/Sidebar.tsx
