+++ packages/toolkit/src/view/catalog/components/WarnDiscardFilesDialog.tsx
