+++ packages/toolkit/src/view/catalog/components/lib/constant.ts
