+++ packages/toolkit/src/view/catalog/components/lib/helpers.tsx
