+++ packages/toolkit/src/view/catalog/components/notifications/DeleteCatalogNotification.tsx
