+++ packages/toolkit/src/view/catalog/components/notifications/InsufficientCreditNotification.tsx
