+++ packages/toolkit/src/view/catalog/components/tabs/ImageTab.tsx
