+++ packages/toolkit/src/view/catalog/components/tabs/KnowledgeBaseTab.tsx
