+++ packages/toolkit/src/view/catalog/components/tabs/index.ts
