+++ packages/toolkit/src/view/index.ts
@@ -5,4 +5,4 @@ export * from "./pipeline";
 export * from "./pipeline-builder";
 export * from "./profile/user-profile-view";
 export * from "./settings";
-export * from "./knowledge";
+export * from "./catalog";
